[VERY ROUGHLY] seperate rails.js into multiple modules
Most tests pass, except for three override tests since we are using a
"static linking" provided by ES2015 now.

It'll be fixed in the following commits.
